fix: make vmware platform common code build on all arches
Even though VMWare doesn't support non-amd64 case, for the imager (and
Image Factory), the common stuff should work correctly for any arch the
imager is running with (as arm64 imager can generate amd64 VMWare
image).

See https://github.com/siderolabs/image-factory/issues/164
