+++ internal/app/machined/pkg/runtime/v1alpha1/platform/vmware/vmware_other.go
@@ -11,32 +11,13 @@ import (
 	"errors"
 
 	"github.com/cosi-project/runtime/pkg/state"
-	"github.com/siderolabs/go-procfs/procfs"
 
 	"github.com/siderolabs/talos/internal/app/machined/pkg/runtime"
 )
 
-// VMware is the concrete type that implements the platform.Platform interface.
-type VMware struct{}
-
-// Name implements the platform.Platform interface.
-func (v *VMware) Name() string {
-	return "vmware"
-}
-
 // Configuration implements the platform.Platform interface.
 func (v *VMware) Configuration(context.Context, state.State) ([]byte, error) {
 	return nil, errors.New("arch not supported")
-}
-
-// Mode implements the platform.Platform interface.
-func (v *VMware) Mode() runtime.Mode {
-	return runtime.ModeCloud
-}
-
-// KernelArgs implements the runtime.Platform interface.
-func (v *VMware) KernelArgs(string) procfs.Parameters {
-	return []*procfs.Parameter{}
 }
 
 // NetworkConfiguration implements the runtime.Platform interface.
